UI, API, logic

+++ apps/ios/SimpleXChat/ChatTypes.swift
@@ -3385,7 +3385,7 @@ public struct FileTransferMeta: Decodable {
     public let fileId: Int64
     public let fileName: String
     public let filePath: String
-    public let fileSize: Int
+    public let fileSize: Int64
 }
 
 public enum CICallStatus: String, Decodable {
